fix(android): Restore signed id schema for JavaClass method list (#1675)
get_java_method_list() cast the opaque JNI method identifier straight
to uint64_t, which after routing unsigned C++ integer constructors to
Variant::UINT started exposing "id" as TYPE_UINT. The method is
documented to match Object.get_method_list()'s signed carrier, so
round-trip the identifier through uint64_t and back to int64_t to
keep its bit pattern while landing on the correct Variant type.

Extends the instrumented JavaClassWrapper Foundry Script test to pull
a nonempty method list and assert every id is TYPE_INT.

+++ platform/android/java_class_wrapper.cpp
@@ -756,7 +756,12 @@ TypedArray<Dictionary> JavaClass::get_java_method_list() const {
 			Dictionary method;
 
 			method["name"] = mi._constructor ? java_constructor_name : String(item.key);
-			method["id"] = (uint64_t)mi.method;
+			// mi.method is an opaque JNI method identifier; the public method-list
+			// dictionary must still follow Object.get_method_list()'s signed "id"
+			// carrier, so round-trip through uint64_t to preserve the bit pattern
+			// and land on the signed Variant type. Do not simplify to a direct
+			// uint64_t assignment, which would expose TYPE_UINT here instead.
+			method["id"] = int64_t(uint64_t(mi.method));
 			method["default_args"] = Array();
 			method["flags"] = METHOD_FLAGS_DEFAULT & (mi._static || mi._constructor ? METHOD_FLAG_STATIC : METHOD_FLAG_NORMAL);
